+++ src/AnalyticsFacade.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Spatie\Analytics;
+
+use Illuminate\Support\Facades\Facade;
+
+/**
+ * @see \Spatie\Analytics\Analytics
+ */
+class AnalyticsFacade extends Facade
+{
+    protected static function getFacadeAccessor()
+    {
+        return 'laravel-analytics';
+    }
+}