+++ tests/Integration/TestCase.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace Spatie\Analytics\Tests\Integration;
+
+use Spatie\Analytics\AnalyticsFacade;
+use Orchestra\Testbench\TestCase as Orchestra;
+use Spatie\Analytics\AnalyticsServiceProvider;
+
+abstract class TestCase extends Orchestra
+{
+    public function setUp()
+    {
+        parent::setUp();
+    }
+
+    /**
+     * @param \Illuminate\Foundation\Application $app
+     *
+     * @return array
+     */
+    protected function getPackageProviders($app)
+    {
+        return [
+            AnalyticsServiceProvider::class,
+        ];
+    }
+
+    /**
+     * @param \Illuminate\Foundation\Application $app
+     *
+     * @return array
+     */
+    protected function getPackageAliases($app)
+    {
+        return [
+            'Analytics' => AnalyticsFacade::class,
+        ];
+    }
+}