+++ CHANGELOG.md
@@ -1,6 +1,63 @@
 # Changelog
 
-All Notable changes to Laravel-Analytics will be documented in this file
+All notable changes to Laravel-Analytics will be documented in this file
+
+## 3.0.0 - 2016-08-23
+- add support for `v2` of the Google API
+- renamed config file from `laravel-analytics.php` to `analytics.php`
+- dropped support for anything lower than Laravel 5.4
+
+## 2.4.0 - 2017-01-23
+- add support for Laravel 5.4
+- dropped support for anything lower than Laravel 5.3
+
+## 2.3.1 - 2016-10-14
+- improve exception message
+
+## 2.3.0 - 2016-10-20
+- added `fetchTotalVisitorsAndPageViews`
+
+## 2.2.2 - 2016-08-23
+- added L5.3 compatibility
+
+## 2.2.1 - 2016-08-02
+- added a fallback for the cache path setting
+
+## 2.2.0 - 2016-07-23
+- added config setting to specify cache path
+
+## 2.1.0 - 2016-06-22
+- added `pageTitle` to `fetchVisitorsAndPageViews` and `fetchMostVisitedPages`
+- fixed `credentialsJsonDoesNotExist` exception
+
+## 2.0.0 - 2016-06-20
+
+- refactored all methods
+- introduced `Spatie\Analytics\Period` to specify date ranges
+- the package now uses json credentials instead of .p12 file
+- the `Spatie\Analytics\Analtyics` class is much easier to extend
+
+## 1.4.1 - 2016-05-12
+
+- fixes a bug introduced in 1.4.0 where the check whether a p12 certificate is present was broken 
+
+## 1.4.0 - 2016-05-11
+
+**This version does not work, please upgrade to 1.4.0**
+
+- removing the use of facades
+
+## 1.3.1
+- make `calculateNumberOfDays`-function protected
+
+## 1.3.0
+- Added method getSiteId() 
+
+## 1.2.3
+- Fix PHP 5.4 compatibility
+
+## 1.2.2
+- An injected `Spatie\LaravelAnalytics\LaravelAnalytics`-object will now be properly configured
 
 ## 1.2.1
 - Removed a var_dump-call that was not supposed to be there